Add files via upload
This problem simulates interactions among dierent forms of life in a world. The world is represented by an N  N grid that changes over a number of cycles. Within a cycle, each square is occupied by one of the following life life forms:
Badger (B), Fox (F), Rabbit (R), Grass (G), and Empty (E)
An Empty square means that it is not occupied by any life form.

The world evolves from one cycle to the next one. In the next cycle, the life form residing on every square is decided from that in the current cycle as well as those in the 3x3 neighborhood centered at the square, according to some survival rules. 
Badgers, Foxes, and Rabbits start at age 0 and grow one year older when the next cycle starts.
 For every life form, a set of rules apply.